deleted .env updated utils and bot.py
