+++ .gitignore
@@ -6,6 +6,7 @@
 app/__pycache__/
 app/.env
 
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[codz]
