--- app/.env
@@ -1,1 +1,0 @@
-BOT_TOKEN=7579539691:AAE5Q0gn7J_VocrrV1xAfSIHJuJ3Akz0H4E
