+++ app/bot.py
@@ -7,6 +7,8 @@ import re
 from api import get_schedule
 from groupes import groups
 from loguru import logger
+import hashlib
+
 
 # aiogram imports
 from aiogram import Bot, Dispatcher, html
@@ -14,6 +16,7 @@ from aiogram.client.default import DefaultBotProperties
 from aiogram.enums import ParseMode
 from aiogram.filters import CommandStart
 from aiogram.types import Message
+from aiogram import types
 
 # from utils
 from utils import get_inline_keyboard_select_group
@@ -53,6 +56,58 @@ async def handler(message: Message):
                 #reply_markup=get_inline_keyboard()
             )
 
+@dp.callback_query(lambda c: c.data == "search")
+async def process_search(callback_query):
+    await callback_query.message.answer("Please enter the group code (e.g., АП-11):")
+    await callback_query.answer()
+
+@dp.callback_query(lambda c: c.data == "button1")
+async def process_button1(callback_query):
+    await callback_query.message.answer("You clicked Button 1!")
+    await callback_query.answer()
+    
+@dp.callback_query(lambda c: c.data == "button2")
+async def process_button2(callback_query):
+    await callback_query.message.answer("You clicked Button 2!")
+    await callback_query.answer()
+
+@dp.inline_query()
+async def inline_handler(inline_query: types.InlineQuery):
+    query = inline_query.query.strip()
+    logger.info(f"Inline query: '{query}' from user {inline_query.from_user.id}")
+
+    results = []
+
+    if query:  # Чекаем что пользователь В ОБЩЕМ что то ввел 
+        for key, value in groups.items():
+            if query.lower() in key.lower():  #Ищем по подстроке
+                result_id = hashlib.md5(key.encode()).hexdigest()
+                input_content = types.InputTextMessageContent(
+                    message_text=f"Вы выбрали группу: {key} ({value})"
+                )
+                result = types.InlineQueryResultArticle(
+                    id=result_id,
+                    title=f"Группа: {key}",
+                    input_message_content=input_content,
+                    description=f"Код группы: {value}"
+                )
+                results.append(result)
+
+    # Если юзер даун и ввел че то что мы не знаем говорим ему что не найдено 
+    if query and not results:
+        result_id = hashlib.md5(query.encode()).hexdigest()
+        input_content = types.InputTextMessageContent(
+            message_text="Группа не найдена. Пожалуйста, введите корректный код группы."
+        )
+        result = types.InlineQueryResultArticle(
+            id=result_id,
+            title="Группа не найдена",
+            input_message_content=input_content,
+            description="Нет такой группы." #Еле сдержался от мата вхзввххвхв
+        )
+        results.append(result)
+
+    await inline_query.answer(results, cache_time=1)
